feat: login process page route enum
