+++ frontend-hybrid/yakal/lib/screens/Login/KakaoLogin/screen.dart
@@ -6,7 +6,7 @@ import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:get/get.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
-import 'package:yakal/screens/Login/LoginBeforeIdentify/screen.dart';
+import 'package:yakal/screens/Login/LoginProcess/screen.dart';
 import 'package:yakal/viewModels/Profile/user_view_model.dart';
 
 import '../../../utilities/style/color_styles.dart';
@@ -113,13 +113,14 @@ class _KakaoLoginScreenState extends State<KakaoLoginScreen> {
 
       if (userViewModel.user.value.isAgreedMarketing == null) {
         routeController.goToTerms();
-        Get.offNamed("/login/test");
+        Get.offNamed("/login/process");
         return;
       }
 
       final regex = RegExp(r'user#\d{6}');
       if (regex.hasMatch(userViewModel.user.value.nickName)) {
-        Get.offNamed("/login/nickname");
+        routeController.goToInputNickname();
+        Get.offNamed("/login/process");
         return;
       }
 
@@ -150,7 +151,7 @@ class _KakaoLoginScreenState extends State<KakaoLoginScreen> {
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
-      backgroundColor: ColorStyles.white,
+      backgroundColor: Color.fromRGBO(98, 98, 114, 0.4),
       body: Center(
         child: CircularProgressIndicator(
           color: ColorStyles.main,
