+++ frontend-hybrid/yakal/lib/screens/Login/LoginProcess/LoginFinished/screen.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:yakal/screens/Login/LoginFinished/style.dart';
+import 'package:yakal/utilities/enum/login_process.dart';
+import 'package:yakal/utilities/style/color_styles.dart';
+import 'package:yakal/viewModels/Profile/user_view_model.dart';
 import 'package:yakal/widgets/Base/bottom_button.dart';
 import 'package:yakal/widgets/Base/outer_frame.dart';
+import 'package:yakal/widgets/Login/login_app_bar.dart';
 
-import '../../../utilities/enum/login_process.dart';
-import '../../../utilities/style/color_styles.dart';
-import '../../../viewModels/Profile/user_view_model.dart';
-import '../../../widgets/Login/login_app_bar.dart';
+import 'style.dart';
 
 class LoginFinishedScreen extends StatelessWidget {
   const LoginFinishedScreen({super.key});
