+++ frontend-hybrid/yakal/lib/screens/Login/LoginProcess/LoginTerms/TermsDetail/style.dart
