+++ frontend-hybrid/yakal/lib/screens/Login/LoginProcess/LoginTerms/screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:get/get.dart';
-import 'package:yakal/screens/Login/LoginBeforeIdentify/LoginTerms/TermsDetail/screen.dart';
-import 'package:yakal/screens/Login/LoginBeforeIdentify/LoginTerms/style.dart';
-import 'package:yakal/screens/Login/LoginBeforeIdentify/screen.dart';
+import 'package:yakal/screens/Login/LoginProcess/LoginTerms/TermsDetail/screen.dart';
+import 'package:yakal/screens/Login/LoginProcess/LoginTerms/style.dart';
+import 'package:yakal/screens/Login/LoginProcess/screen.dart';
 import 'package:yakal/utilities/style/color_styles.dart';
 import 'package:yakal/viewModels/Profile/user_view_model.dart';
 import 'package:yakal/widgets/Login/auth_check_button.dart';
