+++ frontend-hybrid/yakal/lib/screens/Login/LoginProcess/LoginTerms/style.dart
