+++ frontend-hybrid/yakal/lib/screens/Login/LoginProcess/ModeSelection/screen.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:yakal/screens/Login/ModeSelection/style.dart';
+import 'package:yakal/utilities/enum/login_process.dart';
 import 'package:yakal/utilities/enum/mode.dart';
+import 'package:yakal/utilities/style/color_styles.dart';
 import 'package:yakal/widgets/Base/bottom_button.dart';
 import 'package:yakal/widgets/Base/outer_frame.dart';
+import 'package:yakal/widgets/Login/login_app_bar.dart';
 import 'package:yakal/widgets/Login/mode_selection_box.dart';
 
-import '../../../utilities/enum/login_process.dart';
-import '../../../utilities/style/color_styles.dart';
-import '../../../widgets/Login/login_app_bar.dart';
+import 'style.dart';
 
 class ModeSelectionScreen extends StatefulWidget {
   const ModeSelectionScreen({super.key});
