+++ frontend-hybrid/yakal/lib/screens/Login/LoginProcess/ModeSelection/style.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-
-import '../../../utilities/style/color_styles.dart';
+import 'package:yakal/utilities/style/color_styles.dart';
 
 class ModeSelectionStyle {
   static const selectedModeTitle = TextStyle(
